added various pages

+++ src/pages/edit-note/edit-note.ts
@@ -1,5 +1,5 @@
 import { Component } from '@angular/core';
-import { IonicPage, NavController, NavParams } from 'ionic-angular';
+import { IonicPage, NavController, NavParams, ToastController } from 'ionic-angular';
 import { Note } from '../../model/note/note.model';
 import { NoteListService } from '../../services/note-list.service';
 
@@ -20,11 +20,13 @@ export class EditNotePage {
     pbit: true,
     ibit: true,
     dboi: '',
-    myloc: ''
+    myloc: '',
+    upn: '',
+    dpn: ''
   };
 
   constructor(
-    public navCtrl: NavController,
+    public navCtrl: NavController, private toast: ToastController,
     public navParams: NavParams,
     private noteListService: NoteListService) {
   }
@@ -34,14 +36,55 @@ export class EditNotePage {
   }
 
   updateNote(note: Note) {
-    this.noteListService.updateNote(note).then(() => {
-      this.navCtrl.setRoot('HomePage');
-    })
+    if((note.items==null)||(note.upn==null)||(note.cost==null)){
+      this.toast.create({
+        message: 'Fill all details',
+        duration: 2000
+      }).present();
+    }
+    else{
+      var number = /^[0-9]{10}$/.test(note.upn);
+      if(!number){
+        this.toast.create({
+          message: 'Invalid Contact',
+          duration: 2000
+        }).present();
+      }
+      else{
+        if(this.note.pbit==false){
+          this.toast.create({
+            message: 'Cannot Update, Order already accepted',
+            duration: 2000
+          }).present();
+          this.navCtrl.setRoot('HomePage');
+        }
+        else{
+          this.noteListService.updateNote(note).then(() => {
+            this.navCtrl.setRoot('HomePage');
+          })
+        }
+      }
+    }
+    
+    
   }
 
   removeNote(note: Note) {
-    this.noteListService.removeNote(note).then(() => {
+    if(this.note.pbit==false){
+      this.toast.create({
+        message: 'Cannot Update, Order already accepted',
+        duration: 2000
+      }).present();
       this.navCtrl.setRoot('HomePage');
-    })
+    }
+    else{
+      this.noteListService.removeNote(note).then(() => {
+        this.navCtrl.setRoot('HomePage');
+      })
+    }
+  }
+
+  isReadonly() {
+    return this.isReadonly;   //return true/false 
   }
 }
